Early FlexText integration, changes to base tool & config classes, light documentation, and light formatting changes
-Added parameter to the lqm/tool class, used by all tools, that keeps
track of the tools name
-Added function in the lqm/config class that returns a list of all
'active' tools
-Removed all getters from flextext config; they were redundant
-Added paramater to user config for to_flextext to allow file
incrementing. When set to false a datetime stamp is added to the output
file from to_flextext. Used to prevent overwriting old files
-Processing function in to_flextext has been implemented. The
to_flextext parser is initialized with the to_flextext constructor and
is used in the process function. Current implementation works, but is
using a modified version of flextransform. Will be making a pull request
on flextransform or modifying the ToFlexText class based on feedback.
--Part of this change required some changes to the ToolChain process
method(lqm/tool). Specifically the datafile, which contains the path to
the alerts, is now passed along. This is less efficient, but is
currently the best solution for implementing FlexText.
-Starting to add documentation in the code. Will be going back to older
tools and doing the same soon.
-Minor formatting cleanup, specifically in to_csv/tool.py