FlexText bug fix, Splunk API revisions
- Updated how FlexText configures the `fields` variable. Now uses an
array instead of a string. This fixes a bug where spaces in the user
definition would cause issues when running the tool.
- Documentation has been updated to reflect this configuration change
- Revised the main Splunk API class. Simplified how errors are raised to
used requests standard method
NOTE: these changes are included in the to_bro branch because they would
require a new version of LQMT to be pushed to pypi. Since Bro
implementation is nearly complete, they will be merged with the main
branch with the bro tool so everything can be included in one version
update.